Use container parameter bag instead of own config class
