+++ src/Config.php
@@ -3,6 +3,7 @@
 namespace ShopwareCli;
 
 use ShopwareCli\Services\PathProvider\PathProvider;
+use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -10,100 +11,21 @@ use Symfony\Component\Yaml\Yaml;
  *
  * Class Config
  * @package ShopwareCli
+ * @deprecated Use the container for config handling instead
  */
 class Config implements \ArrayAccess
 {
     /**
-     * @var array
+     * @var \Symfony\Component\DependencyInjection\Container
      */
-    protected $configArray;
-
-    /**
-     * @var PathProvider
-     */
-    private $pathProvider;
+    private $container;
 
     /**
      * @param PathProvider $pathProvider
      */
-    public function __construct(PathProvider $pathProvider)
+    public function __construct(Container $container)
     {
-        $this->pathProvider = $pathProvider;
-
-        $config = $this->getMergedConfigs($this->collectConfigFiles());
-        $this->configArray = Yaml::parse($config, true);
-    }
-
-    /**
-     * Iterate the extension directories and return config.yaml files
-     *
-     * @return string[]
-     */
-    private function collectConfigFiles()
-    {
-        $files = array();
-
-        // Load user file first. Its config values cannot be overwritten
-        $userConfig = $this->pathProvider->getConfigPath() . '/config.yaml';
-        if (file_exists($userConfig)) {
-            $files[] = $userConfig;
-        }
-
-        // load config.yaml files from extensions
-        $vendorIterator = new \DirectoryIterator($this->pathProvider->getExtensionPath());
-        /** @var $vendorFileInfo \DirectoryIterator */
-        foreach ($vendorIterator as $vendorFileInfo) {
-            if (!$this->isValidDir($vendorFileInfo)) {
-                continue;
-            }
-            $extensionIterator = new \DirectoryIterator($vendorFileInfo->getPathname());
-
-            /** @var $extensionFileInfo \DirectoryIterator */
-            foreach ($extensionIterator as $extensionFileInfo) {
-                if (!$this->isValidDir($extensionFileInfo)) {
-                    continue;
-                }
-                $file = $extensionFileInfo->getPathname() . '/config.yaml';
-
-                if (file_exists($file)) {
-                    $files[] = $file;
-                }
-            }
-        }
-
-        // Load config.yaml.dist as latest - this way the fallback config options are defined
-        $files[] = $this->pathProvider->getCliToolPath() . '/config.yaml.dist';
-        return $files;
-    }
-
-    /**
-     * @param \DirectoryIterator $fileInfo
-     *
-     * @return bool
-     */
-    private function isValidDir(\DirectoryIterator $fileInfo)
-    {
-        return $fileInfo->isDir() && !$fileInfo->isDot() && stripos(
-            $fileInfo->getBasename(),
-            '.'
-        ) !== 0; // skip dot directories e.g. .git
-    }
-
-    /**
-     * Merge all given config.yaml files
-     *
-     * @param  string[] $paths
-     * @return string
-     */
-    private function getMergedConfigs($paths)
-    {
-        $content = array();
-
-        foreach ($paths as $path) {
-            $content[] = file_get_contents($path);
-        }
-
-        return implode("\n", $content);
+        $this->container = $container;
     }
 
     /**
@@ -111,7 +33,7 @@ class Config implements \ArrayAccess
      */
     public function getRepositories()
     {
-        return $this->configArray['repositories'];
+        return $this->container->getParameter('repositories');
     }
 
     /**
@@ -121,7 +43,8 @@ class Config implements \ArrayAccess
      */
     public function offsetExists($offset)
     {
-        return isset($this->configArray[$offset]);
+        $var = $this->container->getParameter($offset);
+        return isset($var);
     }
 
     /**
@@ -131,7 +54,7 @@ class Config implements \ArrayAccess
      */
     public function offsetGet($offset)
     {
-        return $this->configArray[$offset];
+        return $this->container->getParameter($offset);
     }
 
     /**
@@ -140,14 +63,15 @@ class Config implements \ArrayAccess
      */
     public function offsetSet($offset, $value)
     {
-        $this->configArray[$offset] = $value;
+        $this->container->setParameter($offset, $value);
     }
 
     /**
      * @param mixed $offset
+     * @throws \RuntimeException
      */
     public function offsetUnset($offset)
     {
-        unset($this->configArray[$offset]);
+        throw new \RuntimeException("Not implemented");
     }
 }
